requirements.txt
numpy>=1.24
pandas>=2.0
matplotlib>=3.7
seaborn>=0.12
scikit-learn>=1.3
tensorflow-cpu>=2.15
yfinance>=0.2.30
xgboost>=2.0
vaderSentiment>=3.3
beautifulsoup4>=4.12
requests>=2.31
streamlit>=1.30
plotly>=5.18
fastapi>=0.100.0
uvicorn>=0.23.0
python-multipart>=0.0.6
